+++ tt_metal/hw/ckernels/wormhole_b0/metal/llk_api/llk_sfpu/llk_math_eltwise_ternary_sfpu_addcmul.h
@@ -18,7 +18,7 @@ inline void llk_math_eltwise_ternary_sfpu_addcmul(
         dst_index1,
         dst_index2,
         odst,
-        static_cast<int>(vector_mode),
+        vector_mode,
         value);
 }
